move typed-redux into app-state
